#1845, move CommandResponsePackets to common
