+++ sharding-proxy/src/main/java/org/apache/shardingsphere/shardingproxy/backend/communication/DatabaseCommunicationEngine.java
@@ -18,7 +18,7 @@
 package org.apache.shardingsphere.shardingproxy.backend.communication;
 
 import org.apache.shardingsphere.shardingproxy.backend.ResultPacket;
-import org.apache.shardingsphere.shardingproxy.transport.mysql.packet.command.CommandResponsePackets;
+import org.apache.shardingsphere.shardingproxy.transport.common.packet.command.CommandResponsePackets;
 
 import java.sql.SQLException;
