+++ sharding-proxy/src/main/java/org/apache/shardingsphere/shardingproxy/backend/text/transaction/SkipBackendHandler.java
@@ -20,7 +20,7 @@ package org.apache.shardingsphere.shardingproxy.backend.text.transaction;
 import org.apache.shardingsphere.shardingproxy.backend.ResultPacket;
 import org.apache.shardingsphere.shardingproxy.backend.text.TextProtocolBackendHandler;
 import org.apache.shardingsphere.shardingproxy.transport.common.packet.generic.DatabaseSuccessPacket;
-import org.apache.shardingsphere.shardingproxy.transport.mysql.packet.command.CommandResponsePackets;
+import org.apache.shardingsphere.shardingproxy.transport.common.packet.command.CommandResponsePackets;
 
 /**
  * Skip backend handler.
