+++ sharding-proxy/src/test/java/org/apache/shardingsphere/shardingproxy/backend/sctl/ShardingCTLSetBackendHandlerTest.java
@@ -18,7 +18,7 @@
 package org.apache.shardingsphere.shardingproxy.backend.sctl;
 
 import org.apache.shardingsphere.shardingproxy.backend.communication.jdbc.connection.BackendConnection;
-import org.apache.shardingsphere.shardingproxy.transport.mysql.packet.command.CommandResponsePackets;
+import org.apache.shardingsphere.shardingproxy.transport.common.packet.command.CommandResponsePackets;
 import org.apache.shardingsphere.shardingproxy.transport.mysql.packet.generic.ErrPacket;
 import org.apache.shardingsphere.shardingproxy.transport.mysql.packet.generic.OKPacket;
 import org.apache.shardingsphere.transaction.core.TransactionType;
