+++ sharding-proxy/src/test/java/org/apache/shardingsphere/shardingproxy/backend/text/transaction/SkipBackendHandlerTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.shardingproxy.backend.text.transaction;
 
-import org.apache.shardingsphere.shardingproxy.transport.mysql.packet.command.CommandResponsePackets;
+import org.apache.shardingsphere.shardingproxy.transport.common.packet.command.CommandResponsePackets;
 import org.apache.shardingsphere.shardingproxy.transport.mysql.packet.generic.OKPacket;
 import org.junit.Test;
